Moved Error enums into their own files in the Error Handling group. Also adjusted some print statemetns for when the sinks to get photo data fail.
