+++ UOKiosk/Events/Models/NonCoreDataModels/IMEvent.swift
@@ -73,10 +73,9 @@ class IMEvent: Identifiable {
                 // TODO: Add error handling
                 switch completion {
                 case .finished:
-                    //print("SUCCEEDED to get photo for event: \(self.title)")
                     break
                 case .failure(let error):
-                    print("FAILED to get photo for event: \(self.title)\n With error: \(error.localizedDescription)")
+                    print("!!! Failed to get photo for event, \(self.title), with error: \(error.localizedDescription)")
                 }
                 self.cancellables.removeAll()
             } receiveValue: { [weak self] (data, respose) in
